requirements.txt
opencv-python
langchain
langchain-openai
langchain_google_genai
langchain-community
python-dotenv
pyaudio
soundfile
openai
numpy
Pillow
openai-whisper
SpeechRecognition
SpeechRecognition[whisper-local]

